+++ apps/endringsmelding-pleiepenger/src/app/modules/lovbestemt-ferie-modal/LovbestemtFerieModal.tsx
@@ -1,8 +1,7 @@
-import { Button, Heading, Modal } from '@navikt/ds-react';
-import React, { FunctionComponent, useEffect } from 'react';
-import { Close } from '@navikt/ds-icons';
-import Block from '@navikt/sif-common-core-ds/lib/atoms/block/Block';
+import { Modal } from '@navikt/ds-react';
+import React, { FunctionComponent } from 'react';
 import './lovbestemtFerieModal.css';
+import { createPortal } from 'react-dom';
 
 interface Props {
     title: string;
@@ -12,40 +11,26 @@ interface Props {
 }
 
 const LovbestemtFerieModal: FunctionComponent<Props> = ({ children, title, open = false, onClose }) => {
-    useEffect(() => {
-        Modal.setAppElement('#app');
-    });
-    return (
-        <Modal
-            open={open}
-            onClose={onClose}
-            className="lovbestemtFerieModal"
-            closeButton={false}
-            aria-label="Lovbestemt ferie">
-            <Button
-                className="navds-modal__button"
-                size="small"
-                variant="tertiary"
-                onClick={onClose}
-                tabIndex={0}
-                icon={<Close title="Lukk modalvindu" aria-label="Lukk modalvindu" />}
-            />
-
-            <Modal.Content>
-                <div style={{ marginTop: 'var(--a-spacing-1)', paddingBottom: 'var(--a-spacing-2)' }}>
-                    <Heading
-                        spacing={true}
-                        size="small"
-                        level="1"
-                        id="lovbestemtFerieModalHeader"
-                        className="lovbestemtFerieModal__noFocusOutline">
-                        {title}
-                    </Heading>
-                    <Block margin="l">{children}</Block>
-                </div>
-            </Modal.Content>
-        </Modal>
-    );
+    return open
+        ? createPortal(
+              <Modal
+                  open={open}
+                  onClose={onClose}
+                  className="lovbestemtFerieModal"
+                  aria-label="Lovbestemt ferie"
+                  header={{
+                      heading: title,
+                      closeButton: true,
+                  }}>
+                  <Modal.Body>
+                      <div style={{ marginTop: 'var(--a-spacing-1)', paddingBottom: 'var(--a-spacing-2)' }}>
+                          {children}
+                      </div>
+                  </Modal.Body>
+              </Modal>,
+              document.body
+          )
+        : null;
 };
 
 export default LovbestemtFerieModal;