+++ apps/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @navikt/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn
 
+## 1.14.0
+
+### Minor Changes
+
+-   026fc373: Oppdatere til versjon 5 av designsystemet aksel
+
+### Patch Changes
+
+-   Updated dependencies [41d0a03c]
+-   Updated dependencies [026fc373]
+    -   @navikt/sif-common-formik-ds@0.3.0
+    -   @navikt/sif-common-soknad-ds@10.3.0
+    -   @navikt/sif-common-core-ds@9.3.0
+
 ## 1.13.0
 
 ### Minor Changes
@@ -12,8 +26,8 @@
 -   Updated dependencies
     -   @navikt/sif-common-amplitude@2.19.0
     -   @navikt/sif-common-formik-ds@0.24.0
-    -   @navikt/sif-common-soknad-ds@11.0.0
-    -   @navikt/sif-common-core-ds@10.0.0
+    -   @navikt/sif-common-soknad-ds@10.3.0
+    -   @navikt/sif-common-core-ds@9.3.0
     -   @navikt/sif-common-sentry@0.20.0
     -   @navikt/sif-common-utils@3.38.0
 
@@ -60,8 +74,8 @@
 
 -   Updated dependencies [ea79f31c]
     -   @navikt/sif-common-formik-ds@0.23.0
-    -   @navikt/sif-common-soknad-ds@11.0.0
-    -   @navikt/sif-common-core-ds@10.0.0
+    -   @navikt/sif-common-soknad-ds@10.3.0
+    -   @navikt/sif-common-core-ds@9.3.0
     -   @navikt/sif-common-sentry@0.19.0
 
 ## 1.11.0